Implemented first page of opcodes
Implemented all first page opcodes except for interrupts and the prefix codes, which require the second page.

Also started to implement MBC1

Status: Currently am able to run tetris in a loop until it tries to write into ROM due to the HL- load op. My assumption
is that the game is relying on timer interrupts to trigger the animations it shows on the main screen after a time