VRAM viewer added
Added a secondary window that visualizes VRAM (all looks good).
Also implemented and fixed some arithmetic functions like SCF,RL/A

Something seems off with the stack, since both the call/ret test and
the stack test are failing, I'm suspicious that this is in turn causing
most of the other tests to fail.